% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_ger.R
\name{as.numeric_ger}
\alias{as.numeric_ger}
\title{Converts german type to english}
\usage{
as.numeric_ger(x)
}
\description{
Converts german type to english
}
\references{
Converts german style to english style
}
