% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark processing time for typical analytics tasks}
\usage{
benchmark(rows = 1e+05, csv = FALSE, size = "rows", bytes = (3 *
  1024)^2, rep_times = 10)
}
\arguments{
\item{rows}{Number of rows of the testing dataset if `size = "rows"`.}

\item{csv}{If `TRUE`, the csv-functions `read.csv2` and `write.csv`
are included in the benchmarking.}

\item{size}{How you want to specify the size of your
testing dataset. Either "rows" or "bytes". "rows" by default.}

\item{bytes}{Size of the dataset if `size = "bytes"`.
E.g. (3*1024)^2 for 3 MB or 1024^4 for one GB.}

\item{rep_times}{Number of time each benchmarked function is repeated
for the benchmarking.}
}
\description{
Benchmark processing time for typical analytics tasks
}
\examples{
benchmark(csv = T) # include benchmarking of read.csv and write.csv2

# specify the size of the dataset you want to test your system with:
benchmark(size = "bytes", bytes = (5*1024)^2) # in bytes
benchmark(rows = 100000) # in rows

benchmark(rep_times = 20) # specify how many times you want to repeat
# benchmarking one function
}
