% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{as_radians}
\alias{as_radians}
\alias{calc_dist}
\title{Distance in km. Input: data.frames with the column fr$lat, fr$lon, to$lat, to$lon}
\usage{
as_radians(theta = 0)

calc_dist(fr, to)
}
\arguments{
\item{fr}{dataframe of origin location(s), needs following columns: lat, lon}

\item{to}{dataframe of destination location(s), needs following columns: lat, lon}
}
\description{
Distance in km. Input: data.frames with the column fr$lat, fr$lon, to$lat, to$lon
}
\references{
Calculate distance in km
}
