% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ger.R
\name{clean_ger}
\alias{clean_ger}
\title{Convert messy german characters to readable output}
\usage{
clean_ger(x)
}
\description{
Convert messy german characters to readable output
}
\references{
Convert messy german characters to readable output
}
