% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_google_places.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Crawl google places for location types}
\format{An object of class \code{environment} of length 1.}
\usage{
pkg.env
}
\arguments{
\item{store_locations}{store-dataframe for which you crawl locations, with following structure: store_id, lat, lon}

\item{radius}{radius of the crawl in meters, default = 500m}

\item{api_key}{your google places api key}
}
\description{
Crawl google places for location types
}
\references{
Crawl google places for location types
}
\keyword{datasets}
