% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_xlsx.R
\name{df_to_xlsx}
\alias{df_to_xlsx}
\title{Create openxlsx workbook object out of data frames}
\usage{
df_to_xlsx(sheet_name = "Sheet1", df, ...)
}
\arguments{
\item{sheet_name}{name of worksheet}

\item{df}{first data frame in the input list}

\item{...}{more data frames}
}
\description{
Create openxlsx workbook object out of data frames
}
\references{
Create openxlsx workbook object out of data frames
}
