% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_eyp.R
\name{eyp_pal}
\alias{eyp_pal}
\alias{scale_colour_eyp}
\alias{scale_color_eyp}
\alias{scale_fill_eyp}
\title{EYP Color Palette (Discrete) and Scales}
\usage{
eyp_pal()

scale_colour_eyp(...)

scale_color_eyp(...)

scale_fill_eyp(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
An 8-color eyp safe qualitative discrete palette.
}
\references{
Umrechnung EYP-Farbschema in Hex-Werte.pptx
}
\seealso{
Other colour: \code{\link{occ_pal}}
}
