% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes_data.R
\docType{data}
\name{ggthemes_data}
\alias{ggthemes_data}
\title{Palette data for the rlytics package}
\format{A \code{list}.}
\usage{
ggthemes_data
}
\description{
Data used by the palettes in the rlytics package.
}
\keyword{datasets}
