% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_list.R
\name{named_list}
\alias{named_list}
\title{Function to create names for dataframes inside a list}
\usage{
named_list(...)
}
\description{
Function to create names for dataframes inside a list
}
\references{
Function to create names for dataframes inside a list
}
