% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_occ.R
\name{occ_pal}
\alias{occ_pal}
\alias{scale_colour_occ}
\alias{scale_color_occ}
\alias{scale_fill_occ}
\title{occ Color Palette (Discrete) and Scales}
\usage{
occ_pal()

scale_colour_occ(...)

scale_color_occ(...)

scale_fill_occ(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
An 8-color occ safe qualitative discrete palette.
}
\references{
Umrechnung OCC-Farbschema in Hex-Werte.pptx
}
\seealso{
Other colour: \code{\link{eyp_pal}}
}
