% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properFilename.R
\name{properFilename}
\alias{properFilename}
\title{Creates proper filenames in eyp style}
\usage{
properFilename(name = "test", project = "51103", path = "",
  kuerzel = "JSA", fileType = "xlsx")
}
\description{
Creates proper filenames in eyp style
}
\references{

}
