% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_clip2.R
\name{write.clip2}
\alias{write.clip2}
\title{Writes data to clipboard}
\usage{
write.clip2(x, dec = ".", col.names = TRUE, row.names = FALSE)
}
\arguments{
\item{dec}{decimal seperator, default is "."}

\item{row.names}{row.names = FALSE is default}

\item{column.names}{column.names = TRUE is default}
}
\description{
Writes data to clipboard
}
\references{
Writes data to clipboard, with "." as dec
}
