% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{as_radians}
\alias{as_radians}
\alias{calc_dist}
\title{Calculate distance besteen geo coded locations in kilometre}
\usage{
as_radians(theta = 0)

calc_dist(fr, to)
}
\description{
Calculate distance besteen geo coded locations in kilometre
}
\references{
Calculate distance in km
}
